First version with complete working routes. - Improved route object and related extension methods. - Improved get shape to get reverse shapes if needed. - Added functionality to get a shape between two offsets. - Added location on network for a snappoint. - Improved the default routing to return a route instead of a path. - Implemented a route builder. - Append and prepend without offsets and refined offset definition. - Fixed dijkstra path builder. - Improved snapping. - Added position along line calculations. - Added extension method in tests to debug snap points.

+++ src/Itinero/Profiles/RouterDbExtensions.cs
@@ -10,10 +10,10 @@ namespace Itinero.Profiles
             
             if (forward)
             {
-                return length * factor.FactorForward;
+                return (uint)(length * factor.FactorForward);
             }
 
-            return length * factor.FactorBackward;
+            return (uint)(length * factor.FactorBackward);
         }
     }
 }
